Renamed Piece Spawner & Piece Manager
Piece Spawner is now Piece Manager as that is the bettter suited name.
Piece Manager is now Piece Click as that is better suited.
