+++ Honours Project/Assets/Scripts/BoxClick.cs
@@ -11,9 +11,9 @@ public class BoxClick : MonoBehaviour {
 	public void boxEnter(){
 		if(!buttonPressed){
 			defaultColour = GetComponent<Image>().color;
-			// Debug.Log(PieceSpawner.pieceArray[PieceSpawner.instance.returnIndex()].name);
+			// Debug.Log(PieceManager.pieceArray[PieceManager.instance.returnIndex()].name);
 			//Checks if The Piece has been clicked and will place it on the grid if it has. 
-			if (PieceSpawner.instance.selected && GetComponentInChildren<Text>().text == ""){
+			if (PieceManager.instance.selected && GetComponentInChildren<Text>().text == ""){
 				GetComponent<Image>().color = Color.cyan;
 				string objectname = this.name;
 				Debug.Log(objectname + " has been clicked.");
@@ -23,8 +23,8 @@ public class BoxClick : MonoBehaviour {
 
 				if (ValidationManager.PositioningValidation(row, column))
 				{
-					// if(PieceSpawner.instance.firstmove){
-					// 	PieceSpawner.instance.firstmove = false;
+					// if(PieceManager.instance.firstmove){
+					// 	PieceManager.instance.firstmove = false;
 					// 	addPiece();
 					// 	addScore(row,column);  
 					// }else 
@@ -32,19 +32,19 @@ public class BoxClick : MonoBehaviour {
 						addPiece(row, column); 
 						addScore(row,column);
 					} else {
-						PieceSpawner.instance.selected = false; 
+						PieceManager.instance.selected = false; 
 						ErrorManagement.instance.ShowError("Error: Please ensure that the total value is an odd number");
 					}
 
 				} else{
 					ErrorManagement.instance.ShowError("Error: Piece must be placed next to an existing piece.");
 				}
-			} else if (PieceSpawner.instance.selected && GetComponentInChildren<Text>().text != "") {
-				PieceSpawner.instance.selected = false; 
+			} else if (PieceManager.instance.selected && GetComponentInChildren<Text>().text != "") {
+				PieceManager.instance.selected = false; 
 				ErrorManagement.instance.ShowError("Error: Piece cannot be placed ontop of an existing piece.");
-			} else if (!PieceSpawner.instance.selected) {
-				Debug.Log(PieceSpawner.pieceArray[PieceSpawner.instance.returnIndex()].name + "is " + PieceSpawner.instance.selected);
-				PieceSpawner.instance.selected = false; 
+			} else if (!PieceManager.instance.selected) {
+				Debug.Log(PieceManager.pieceArray[PieceManager.instance.returnIndex()].name + "is " + PieceManager.instance.selected);
+				PieceManager.instance.selected = false; 
 				ErrorManagement.instance.ShowError("Error: Please select a piece before placing on the grid.");
 			
 		}
@@ -65,18 +65,18 @@ public class BoxClick : MonoBehaviour {
 
 
 	void addPiece(int row, int column){
-		BoxSpawner.gridArray[row,column].GetComponentInChildren<Text>().text = PieceSpawner.instance.returnPieceValue().ToString();
+		BoxSpawner.gridArray[row,column].GetComponentInChildren<Text>().text = PieceManager.instance.returnPieceValue().ToString();
 
-		PieceSpawner.instance.pieceClicked(PieceSpawner.instance.returnIndex()); 
+		PieceManager.instance.pieceClicked(PieceManager.instance.returnIndex()); 
 		GetComponent<Collider2D>().enabled = false;
-		PieceSpawner.instance.setPieceValue(PieceSpawner.instance.returnIndex());
+		PieceManager.instance.setPieceValue(PieceManager.instance.returnIndex());
 	}
 
 	void addScore(int row, int column){
 		int total = 0;
 		int secondtotal = 0; 
 
-		if (ValidationManager.RowTotal(row, column) != PieceSpawner.instance.returnPieceValue()){
+		if (ValidationManager.RowTotal(row, column) != PieceManager.instance.returnPieceValue()){
 			total = ValidationManager.RowTotal(row, column);
 			// Secondary Column Checks
 			if (BoxSpawner.gridArray[row+1,column].GetComponentInChildren<Text>().text !=""
